requirements.txt
albumentations==1.3.1
apturl==0.5.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
contourpy==1.2.0
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
devscripts===2.22.1ubuntu1
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
einops==0.7.0
entmax==1.2
evdev==1.6.1
fasteners==0.14.1
filelock==3.13.1
fonttools==4.46.0
fsspec==2023.12.2
future==0.18.2
gpg==1.16.0
httplib2==0.20.2
huggingface-hub==0.20.2
idna==3.3
imageio==2.33.1
importlib-metadata==4.6.4
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.3.2
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
llfuse==1.3.8
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.8.2
meson==0.61.2
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
munch==4.0.0
netifaces==0.11.0
networkx==3.2.1
numpy~=1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python-headless==4.9.0.80
packaging==23.2
pandas~=2.2.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.2.0
pix2tex==0.1.2
protobuf==3.12.4
ptyprocess==0.7.0
py3dns==3.2.1
pycairo==1.20.1
pycups==2.0.1
pycurl==7.44.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PyQt6-WebEngine==6.6.0
PyQt6-WebEngine-Qt6==6.6.1
pyRFC3339==1.1
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
PySimpleSOAP==1.16.2
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.2
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-debianbts==3.2.0
python-magic==0.4.24
python-xlib==0.33
pytz==2022.1
pyxdg==0.27
PyYAML~=6.0.1
qudida==0.0.4
regex==2023.12.25
reportbug===11.4.1ubuntu1
reportlab==3.6.8
requests==2.25.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn~=1.4.1.post1
scipy~=1.12.0
scour==0.38.2
screeninfo==0.8.1
SecretStorage==3.3.1
shiboken6==6.6.1
six==1.16.0
sympy==1.12
systemd-python==234
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.5.4
tokenizers==0.15.0
torch~=2.2.0
torchinfo==1.8.0
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
ubuntu-dev-tools==0.193+ubuntu4.22.4.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
unidiff==0.5.5
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
x-transformers==0.15.0
xdg==5
xkit==0.0.0
zipp==1.0.0

customtkinter~=5.2.2